refactor(frontend): replace more core/actions imports
